e2e tests: move entity generation to a new pkg
